check for existence

+++ EM_Creator/ma5Wrapper.py
@@ -195,12 +195,17 @@ class MA5Wrapper:
         dirname = bakeryHelpers.dirName ( process, masses )
         origdatfile = "%s/ANA_%s/Output/SAF/CLs_output_summary.dat" % \
                       ( tempdir, dirname )
+        errFree=True
+        if not os.path.exists ( origdatfile ):
+            errFree=False
+            self.error ( "data file %s does not exist!" % origdatfile )
         destdatfile = bakeryHelpers.datFile (  self.ma5results, self.topo, masses, self.sqrts )
         shutil.move ( origdatfile, destdatfile )
         os.chdir ( self.basedir )
-        self.exe ( "rm -rf %s/ma5cmd*" % self.ma5install )
-        self.exe ( "rm -rf %s/recast*" % self.ma5install )
-        self.exe ( "rm -rf %s" % tempdir )
+        #self.exe ( "rm -rf %s/ma5cmd*" % self.ma5install )
+        #self.exe ( "rm -rf %s/recast*" % self.ma5install )
+        if errFree:
+            self.exe ( "rm -rf %s" % tempdir )
         return 0
 
     def exe ( self, cmd, maxLength=100 ):
